deep-research/pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "A research agent that collects and integrates information from YouTube, Reddit, ArXiv, and Medium"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.4",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
deep-research = "deep_research.main:main"
